// src/types/authResponse.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * An explanation about the purpose of this instance.
 */
export type Error = boolean;
/**
 * An explanation about the purpose of this instance.
 */
export type Message = string;
/**
 * An explanation about the purpose of this instance.
 */
export type TheTokenSchema = string;
/**
 * An explanation about the purpose of this instance.
 */
export type TheExpiresAtSchema = number;

/**
 * The root schema comprises the entire JSON document.
 */
export interface AuthResponse {
  error: Error;
  message: Message;
  data: AuthResponseData;
}
/**
 * An explanation about the purpose of this instance.
 */
export interface AuthResponseData {
  token?: TheTokenSchema;
  expiresAt?: TheExpiresAtSchema;
}