// src/types/signInUser.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * An explanation about the purpose of this instance.
 */
export type TheEmailSchema = string;
/**
 * An explanation about the purpose of this instance.
 */
export type ThePasswordSchema = string;

/**
 * The root schema comprises the entire JSON document.
 */
export interface SignInUser {
  email: TheEmailSchema;
  password: ThePasswordSchema;
}